ustal

+++ app/Http/Controllers/CategoryController.php
@@ -37,20 +37,35 @@ class CategoryController extends Controller
 
     public function destroy($id){
         $categories = Categories::find($id);
+        if (!$categories) {
+            throw new ApiException(404, 'Категория не найдена');
+        }
         $categories->delete();
         return response()->json(['message' => 'Категория успешно удалена'], 200);
+
     }
 
     public function update(CategoryUpdateRequest $request, $id)
     {
-        //Проверка существования
+        // Поиск категории по id
         $categories = Categories::find($id);
         if (!$categories) {
             throw new ApiException(404, 'Категория не найдена');
         }
+
+        // Проверяем, есть ли категория с таким именем уже в базе данных
+        $existingProduct = Categories::where('name', $request->input('name'))->first();
+        if ($existingProduct) {
+            throw new ApiException(422, 'Категория с таким именем уже существует');
+        }
+
+        // Обновление атрибутов категории
         $categories->name = $request->input('name');
+
         // Сохранение изменений в базе данных
         $categories->save();
+
+        // Возврат успешного ответа
         return response()->json(['message' => 'Категория успешно обновлена'], 200);
     }
 }
